maint: Add lp_media_to_string to public API
* lib/lp-media.c (lp_media_to_string): New function.
* lib/play.h: Add the newly added function to library API.
* tests/test-lp-media-new.c: Test the newly added function.
* tests/test-lp-scene-new.c: Make stricter.
* lib/lp-scene.c: Syntax cleanup.

+++ lib/lp-scene.c
@@ -1202,7 +1202,7 @@ lp_scene_to_string (lp_Scene *scene)
                          strbool (scene->prop.lockstep),
                          strbool (scene->prop.slave_audio),
                          scene->prop.text,
-                         (guint) scene->prop.text_color,
+                         scene->prop.text_color,
                          scene->prop.text_font);
   g_assert_nonnull (str);
